requirements.txt
#Django
django>=2.1.1
djangorestframework>=3.8.2
django-cors-headers>=2.4.0
drf-yasg>=1.10.2
nltk>=3.3
whitenoise==4.1

#Dependencies need special handling
numpy>=1.15.1
cryptography>=2.3.1
cffi>=1.11.3
pandas>=0.20.3
scipy>=1.0.0
scikit-learn>=0.19.1